refactor: Search Page 상태 관리 로직을 Custom Hook으로 분리
