+++ app/components/list/VideoListItem.module.scss
@@ -13,7 +13,7 @@
     background-color: $color-o-w1;
   }
 
-  // 썸네일 이미지 컨테이ㄴ
+  // 썸네일 이미지 컨테이너
   .thumbnail_img_container {
     border-radius: 10px;
     overflow: hidden;
@@ -22,27 +22,17 @@
     aspect-ratio: 1 / 1;
     display: flex;
     align-items: center;
+    position: relative;
 
-    .video_thumbnail_img {
+    .thumbnail_img {
       object-fit: cover;
       width: 134%;
       height: 134%;
       transition: transform 0.3s ease;
     }
-
-    .thumbnail_img {
-      transition: filter 0.3s ease;
-      object-fit: cover;
-      width: 100%;
-      height: 100%;
-    }
-
-    .video_player {
-      width: 400px;
-      height: 400px;
-      @include flex;
-    }
   }
+
+  // 설명 컨테이너
   .desc_container {
     width: 100%;
     position: relative;
@@ -82,6 +72,7 @@
   }
 }
 
+// 비디오 아이템 컨테이너
 .video_item_container {
   cursor: default;
   @extend .thumbnail_item_container;
@@ -94,6 +85,7 @@
   }
 }
 
+// 플레이리스트 아이템 컨테이너
 .playlist_item_container {
   cursor: default;
   @extend .thumbnail_item_container;
@@ -134,9 +126,7 @@
   width: 100%;
   height: 100%;
   overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+  @include flex;
 }
 
 .video_thumbnail_img {
@@ -145,10 +135,7 @@
   height: 134% !important;
   max-width: none !important;
   max-height: none !important;
-  transform: translate(-50%, -50%);
-  position: absolute;
-  left: 50%;
-  top: 50%;
+  @include position;
 }
 
 .hidden_player {
